Monitoring sets

+++ src/AppBundle/Command/InstallCommand.php
@@ -6,16 +6,17 @@ use AppBundle\Service\LdapService;
 use AppBundle\Document\Application;
 use AppBundle\Service\KibanaService;
 use AppBundle\Service\ApplicationService;
+use AppBundle\Service\SearchGuardService;
 use Doctrine\ODM\MongoDB\DocumentManager;
 use ATS\PaymentBundle\Service\PlanService;
 use AppBundle\Service\ElasticSearchService;
+use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\HttpKernel\KernelInterface;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Doctrine\Common\DataFixtures\Purger\MongoDBPurger;
 use Doctrine\Common\DataFixtures\Executor\MongoDBExecutor;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
-use Symfony\Component\Console\Input\InputOption;
 
 class InstallCommand extends ContainerAwareCommand
 {
@@ -43,43 +44,55 @@ Load default Application Type. Insert template for Kibana and more..'
         $planService = $this->getContainer()->get(PlanService::class);
         /** @var ApplicationService $applicationService */
         $applicationService = $this->getContainer()->get(ApplicationService::class);
+        /** @var SearchGuardService $sgService */
+        $sgService = $this->getContainer()->get(SearchGuardService::class);
+
+        /** @var bool $stripeEnabled */
+        $stripeEnabled = $this->getContainer()->getParameter('stripe_enabled');
 
         /** @var bool $purge */
         $purge = $input->getOption("purge") === true ?: false;
-        $this->display($output, "Deleting Stripe plans");
-        $planService->deleteAllPlans();
+
+        if ($stripeEnabled === true) {
+            $this->display($output, "Deleting Stripe plans");
+            $planService->deleteAllPlans();
+        }
 
         $this->loadFixtures($output, $purge);
+
         $this->display($output, "Creating LDAP entries for demo applications");
         $ldap->createDemoApplicationsEntries();
         $demoApplications = $applicationService->listDemoApplications();
 
+        $this->display($output, "Initializing SearchGuard configuration");
+        $sgService->updateSearchGuardConfig();
+
         $this->display($output, "Initializing ES & Kibana");
         /** @var Application $application */
         foreach ($demoApplications as $application) {
-            $es->deleteIndex("app_" . $application->getUuid());
+            $es->deleteIndex($application->getApplicationIndex());
             $es->createIndexTemplate($application, $applicationService->getActiveApplicationsNames());
-            $es->createAlias($application->getName());
+            $es->createAlias($application);
 
             $kibana->loadIndexPatternForApplication(
                 $application,
-                $application->getOwner(),
-                'app_' . $application->getUuid()
+                $application->getApplicationIndex()
             );
 
-            $kibana->createApplicationDashboards($application, $application->getOwner());
+            $kibana->createApplicationDashboards($application);
 
-            $es->deleteIndex("shared_" . $application->getUuid());
+            $es->deleteIndex($application->getSharedIndex());
 
             $kibana->loadIndexPatternForApplication(
                 $application,
-                $application->getOwner(),
-                'shared_' . $application->getUuid()
+                $application->getSharedIndex()
             );
         }
 
-        $this->display($output, "Creating Stripe Plans with new Data");
-        $planService->createDefaultPlans();
+        if ($stripeEnabled === true) {
+            $this->display($output, "Creating Stripe Plans with new Data");
+            $planService->createDefaultPlans();
+        }
 
         return 0;
     }
